--- src/main/java/urecagroup1backend/orders/api/docs/OrdersControllerDocs.java
@@ -1,2 +1,0 @@
-package urecagroup1backend.orders.api.docs;public interface OrdersControllerDocs {
-}
