--- src/main/java/urecagroup1backend/orders/controller/OrdersController.java
@@ -1,2 +1,0 @@
-package urecagroup1backend.orders.controller;public class OrdersController {
-}
