+++ src/main/java/urecagroup1backend/orders/controller/docs/OrderControllerDocs.java
@@ -1,0 +1,2 @@
+package urecagroup1backend.orders.controller.docs;public interface OrderControllerDocs {
+}
