+++ src/main/java/urecagroup1backend/orders/repository/SnackOrderRepository.java
@@ -1,6 +1,7 @@
 package urecagroup1backend.orders.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import urecagroup1backend.orders.domain.SnackOrder;
 
 public interface SnackOrderRepository extends JpaRepository<SnackOrder, Long> {
 }
